refactor: extract all logic to webworker
